+++ tests/test_indicators.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 # Tests for the Indicator objects
 import gc
+from inspect import _empty
 
 import dask
 import numpy as np
@@ -17,14 +18,14 @@ from xclim.core.formatting import (
     parse_doc,
     update_history,
 )
-from xclim.core.indicator import Indicator, registry
-from xclim.core.missing import missing_pct
+
+from xclim.core.indicator import Daily, Indicator, registry
 from xclim.core.units import units
 from xclim.indices import tg_mean
 from xclim.indices.generic import select_time
 
 
-class UniIndTemp(Indicator):
+class UniIndTemp(Daily):
     realm = "atmos"
     identifier = "tmin"
     var_name = "tmin{thresh}"
@@ -41,7 +42,7 @@ class UniIndTemp(Indicator):
         return out.resample(time=freq).mean(keep_attrs=True)
 
 
-class UniIndPr(Indicator):
+class UniIndPr(Daily):
     realm = "atmos"
     identifier = "prmax"
     units = "mm/s"
@@ -53,7 +54,7 @@ class UniIndPr(Indicator):
         return da.resample(time=freq).mean(keep_attrs=True)
 
 
-class UniClim(Indicator):
+class UniClim(Daily):
     realm = "atmos"
     identifier = "clim"
     units = "K"
@@ -64,7 +65,7 @@ class UniClim(Indicator):
         return select.mean(dim="time", keep_attrs=True)
 
 
-class MultiTemp(Indicator):
+class MultiTemp(Daily):
     realm = "atmos"
     identifier = "minmaxtemp"
     var_name = ["tmin", "tmax"]
@@ -281,9 +282,43 @@ def test_formatting(pr_series):
     ]
 
 
-# TODO Add a meaningful test
 def test_parse_doc():
-    parse_doc(tg_mean.__doc__)
+    doc = parse_doc(tg_mean.__doc__)
+    assert doc["title"] == "Mean of daily average temperature."
+    assert (
+        doc["abstract"]
+        == "Resample the original daily mean temperature series by taking the mean over each period."
+    )
+    assert (
+        doc["parameters"]["tas"]["description"] == "Mean daily temperature [℃] or [K]"
+    )
+    assert (
+        doc["parameters"]["freq"]["description"]
+        == 'Resampling frequency; Defaults to "YS" (yearly).'
+    )
+    assert doc["notes"].startswith("Let")
+    assert "math::" in doc["notes"]
+    assert "references" not in doc
+    assert doc["long_name"] == "The mean daily temperature at the given time frequency"
+
+    doc = parse_doc(xclim.indices.saturation_vapor_pressure.__doc__)
+    assert (
+        doc["parameters"]["ice_thresh"]["description"]
+        == "Threshold temperature under which to switch to equations in reference to ice instead of water. If None (default) everything is computed with reference to water."
+    )
+    assert "Goff, J. A., and S. Gratch (1946)" in doc["references"]
+
+
+def test_parsed_doc():
+    assert "tas" in xclim.atmos.liquid_precip_accumulation.parameters
+    assert "tas" not in xclim.atmos.precip_accumulation.parameters
+
+    params = xclim.atmos.drought_code.parameters
+    assert params["tas"]["description"] == "Noon temperature."
+    assert params["tas"]["annotation"] is xr.DataArray
+    assert params["tas"]["default"] is _empty
+    assert params["snd"]["default"] is None
+    assert params["shut_down_mode"]["annotation"] is str
 
 
 def test_default_formatter():